requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
psycopg2-binary>=2.9
django-netfields>=1.3
python-decouple>=3.8
pytest-django>=4.8
pytest>=8.0
